more i64 -> u64 refactoring

+++ src/extract.rs
@@ -90,10 +90,10 @@ pub struct AFLJFuncDetails {
     pub callrefs: Vec<Callref>,
     #[serde(default)]
     pub datarefs: Vec<DataRef>,
-    pub indegree: Option<i64>,
-    pub outdegree: Option<i64>,
-    pub nlocals: Option<i64>,
-    pub nargs: Option<i64>,
+    pub indegree: Option<u64>,
+    pub outdegree: Option<u64>,
+    pub nlocals: Option<u64>,
+    pub nargs: Option<u64>,
     pub bpvars: Option<Vec<Bpvar>>,
     pub spvars: Option<Vec<Value>>,
     pub regvars: Option<Vec<Regvar>>,
@@ -101,7 +101,7 @@ pub struct AFLJFuncDetails {
     #[serde(default)]
     pub codexrefs: Option<Vec<Codexref>>,
     #[serde(default)]
-    pub dataxrefs: Option<Vec<i64>>,
+    pub dataxrefs: Option<Vec<u64>>,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
@@ -113,10 +113,10 @@ pub struct DataRef {
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Callref {
-    pub addr: i128,
+    pub addr: u64,
     #[serde(rename = "type")]
     pub type_field: String,
-    pub at: i64,
+    pub at: u64,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
@@ -134,7 +134,7 @@ pub struct Bpvar {
 #[serde(rename_all = "camelCase")]
 pub struct Ref {
     pub base: String,
-    pub offset: i64,
+    pub offset: u64,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
@@ -151,10 +151,10 @@ pub struct Regvar {
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Codexref {
-    pub addr: i64,
+    pub addr: u64,
     #[serde(rename = "type")]
     pub type_field: String,
-    pub at: i64,
+    pub at: u64,
 }
 
 // Structs related to AEAFJ
@@ -446,7 +446,7 @@ impl FileToBeProcessed {
 
     fn get_function_xref_details(
         &self,
-        function_addr: i64,
+        function_addr: u64,
         r2p: &mut R2Pipe,
     ) -> Vec<FunctionXrefDetails> {
         info!("Getting function xref details");
@@ -471,7 +471,7 @@ impl FileToBeProcessed {
         json_obj
     }
 
-    fn get_function_info(&self, function_addr: i64, r2p: &mut R2Pipe) -> Vec<AFIJFunctionInfo> {
+    fn get_function_info(&self, function_addr: u64, r2p: &mut R2Pipe) -> Vec<AFIJFunctionInfo> {
         Self::go_to_address(r2p, function_addr);
         let json = r2p.cmd("afij").expect("afij command failed");
         let json_obj: Vec<AFIJFunctionInfo> =
@@ -497,7 +497,7 @@ impl FileToBeProcessed {
         .unwrap_or_else(|_| panic!("the world is ending: {}", f_name));
     }
 
-    fn go_to_address(r2p: &mut R2Pipe, function_addr: i64) {
+    fn go_to_address(r2p: &mut R2Pipe, function_addr: u64) {
         r2p.cmd(format!("s @ {}", function_addr).as_str())
             .expect("failed to seek addr");
     }
